Move FilesystemLockError to 'errors' module
